+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.<br/>
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).<br/><br/>
 
+## [0.3.0] - 2018-10-01
+* updated `package.json` to include the [vue-hot-reload-api](https://www.npmjs.com/package/vue-hot-reload-api) package.
+* updated `Configuration` interface by adding a `hot` property.
+* updated `ASTClass` class by adding the `injectVueHotReloadApi()` method to inject the vue hot reload api business logic.
+
 ## [0.2.0] - 2018-09-29
 * updated `package.json` to include the [sass](https://www.npmjs.com/package/sass) package.